Cleanup my code and add some feature to make it safer
